[mod/#90] api 파일 수정
