+++ data/store/src/main/java/com/napzak/market/store/datasource/StoreDataSource.kt
@@ -1,5 +1,7 @@
 package com.napzak.market.store.datasource
 
+import com.napzak.market.store.dto.StoreDetailResponse
+import com.napzak.market.store.dto.StoreEditProfileResponse
 import com.napzak.market.store.dto.StoreResponse
 import com.napzak.market.store.service.StoreService
 import javax.inject.Inject
@@ -9,5 +11,13 @@ class StoreDataSource @Inject constructor(
 ) {
     suspend fun getStoreInfo(): StoreResponse {
         return storeService.getStoreInfo().data
+    }
+
+    suspend fun getStoreDetail(): StoreDetailResponse {
+        return storeService.getStoreDetail().data
+    }
+
+    suspend fun getEditProfile(): StoreEditProfileResponse {
+        return storeService.getStoreEditProfile().data
     }
 }
