--- data/store/src/main/java/com/napzak/market/store/datasource/StoreDetailDataSource.kt
@@ -1,13 +1,0 @@
-package com.napzak.market.store.datasource
-
-import com.napzak.market.store.dto.StoreDetailResponse
-import com.napzak.market.store.service.StoreDetailService
-import javax.inject.Inject
-
-class StoreDetailDataSource @Inject constructor(
-    private val storeDetailService: StoreDetailService
-) {
-    suspend fun getStoreDetail(): StoreDetailResponse {
-        return storeDetailService.getStoreDetail().data
-    }
-}
