--- data/store/src/main/java/com/napzak/market/store/service/StoreDetailService.kt
@@ -1,12 +1,0 @@
-package com.napzak.market.store.service
-
-import com.napzak.market.remote.model.BaseResponse
-import com.napzak.market.store.dto.StoreDetailResponse
-import retrofit2.http.GET
-
-interface StoreDetailService {
-
-    @GET("api/v1/stores/{storeId}")
-    suspend fun getStoreDetail(
-    ): BaseResponse<StoreDetailResponse>
-}
