--- domain/store/src/main/java/com/napzak/market/store/repository/StoreDetailRepository.kt
@@ -1,7 +1,0 @@
-package com.napzak.market.store.repository
-
-import com.napzak.market.store.model.StoreDetail
-
-interface StoreDetailRepository {
-    suspend fun fetchStoreDetail(token: String): Result<StoreDetail>
-}
