--- domain/store/src/main/java/com/napzak/market/store/repository/StoreEditProfileRepository.kt
@@ -1,7 +1,0 @@
-package com.napzak.market.store.repository
-
-import com.napzak.market.store.model.StoreEditProfile
-
-interface StoreEditProfileRepository {
-    suspend fun fetchEditProfile(token: String): Result<StoreEditProfile>
-}
